# Conflicts:
#	LightMaster/LightMaster/ActionBase.h

+++ LightMaster/LightMaster/ColorManager.cpp
@@ -19,99 +19,99 @@ ColorManager::~ColorManager()
 }
 
 
-int * ColorManager::SetColor(int  size, int  color[])
+int * ColorManager::SetColor(byte  size, byte  color[])
 {
 	int * ret = new int[size];
 
-	for (int a = 0; a < size; a++)
+	for (byte a = 0; a < size; a++)
 		ret[a] = color[a];
 
 	return ret;
 }
 
-int * ColorManager::SetRed(int & size)
+int * ColorManager::SetRed(byte & size)
 {
 
-	int color[] =  { REDPIN };
-	size = sizeof(color) / sizeof(int);
+	byte color[] =  { REDPIN };
+	size = sizeof(color) / sizeof(byte);
 	return SetColor(size, color);
 	
 }
 
-int * ColorManager::SetOrange(int & size)
+int * ColorManager::SetOrange(byte & size)
 {	
-	int color[] = { ORANGEPIN };
-	size = sizeof(color) / sizeof(int);
+	byte color[] = { ORANGEPIN };
+	size = sizeof(color) / sizeof(byte);
 	return SetColor(size, color);	
 }
 
-int * ColorManager::SetYellow(int & size)
+int * ColorManager::SetYellow(byte & size)
 {
-	int color[] = { YELLOWPIN };
-	size = sizeof(color) / sizeof(int);
+	byte color[] = { YELLOWPIN };
+	size = sizeof(color) / sizeof(byte);
 	return SetColor(size, color);	
 }
 
-int * ColorManager::SetGreen(int & size)
+int * ColorManager::SetGreen(byte & size)
 {
-	int color[] = { GREENPIN }; 
-	size = sizeof(color) / sizeof(int);
+	byte color[] = { GREENPIN }; 
+	size = sizeof(color) / sizeof(byte);
 	return SetColor(size, color);	
 }
 
-int * ColorManager::SetRedOrange(int & size)
+int * ColorManager::SetRedOrange(byte & size)
 {
-	int color[] = { REDPIN, ORANGEPIN };
-	size = sizeof(color) / sizeof(int);
+	byte color[] = { REDPIN, ORANGEPIN };
+	size = sizeof(color) / sizeof(byte);
 	return SetColor(size, color);
 }
 
-int * ColorManager::SetRedYellow(int & size)
+int * ColorManager::SetRedYellow(byte & size)
 {
-	int color[] = { REDPIN, YELLOWPIN };
-	size = sizeof(color) / sizeof(int);
+	byte color[] = { REDPIN, YELLOWPIN };
+	size = sizeof(color) / sizeof(byte);
 	return SetColor(size, color);
 }
 
-int * ColorManager::SetRedGreen(int & size)
+int * ColorManager::SetRedGreen(byte & size)
 {
-	int color[] = { REDPIN   , GREENPIN };
-	size = sizeof(color) / sizeof(int);
+	byte color[] = { REDPIN   , GREENPIN };
+	size = sizeof(color) / sizeof(byte);
 	return SetColor(size, color);
 }
 
 //
-int * ColorManager::SetOrangeYellow(int & size)
+int * ColorManager::SetOrangeYellow(byte & size)
 {
-	int color[] = { ORANGEPIN   , YELLOWPIN };
-	size = sizeof(color) / sizeof(int);
+	byte color[] = { ORANGEPIN   , YELLOWPIN };
+	size = sizeof(color) / sizeof(byte);
 	return SetColor(size, color);
 }
-int * ColorManager::SetOrangeGreen(int & size)
+int * ColorManager::SetOrangeGreen(byte & size)
 {
-	int color[] = { ORANGEPIN   , GREENPIN };
-	size = sizeof(color) / sizeof(int);
+	byte color[] = { ORANGEPIN   , GREENPIN };
+	size = sizeof(color) / sizeof(byte);
 	return SetColor(size, color);
 }
-int * ColorManager::SetYellowGreen(int & size)
+int * ColorManager::SetYellowGreen(byte & size)
 {
-	int color[] = { YELLOWPIN   , GREENPIN };
-	size = sizeof(color) / sizeof(int);
+	byte color[] = { YELLOWPIN   , GREENPIN };
+	size = sizeof(color) / sizeof(byte);
 	return SetColor(size, color);
 }
-int * ColorManager::SetOrangeRedYellowOrange(int & size)
+int * ColorManager::SetOrangeRedYellowOrange(byte & size)
 {
-	int color[] = { REDPIN, ORANGEPIN , YELLOWPIN };
-	size = sizeof(color) / sizeof(int);
+	byte color[] = { REDPIN, ORANGEPIN , YELLOWPIN };
+	size = sizeof(color) / sizeof(byte);
 	return SetColor(size, color);
 }
 
 //
 
-int * ColorManager::SetRedOrangeYellowGreen(int & size)
+int * ColorManager::SetRedOrangeYellowGreen(byte & size)
 {
-	int color[] = { REDPIN, ORANGEPIN , YELLOWPIN ,GREENPIN };
-	size = sizeof(color) / sizeof(int);
+	byte color[] = { REDPIN, ORANGEPIN , YELLOWPIN ,GREENPIN };
+	size = sizeof(color) / sizeof(byte);
 	return SetColor(size, color);
 }
